Update types for jsonwebtoken verify()
The `verify()` method accepts a `null` secret when the JWT being verified has `alg: 'none'` (meaning it is unsigned).  The only way to verify an unsigned token is to pass `null` as the secret, as passing an invalid secret results in an error.

The library allows this but the types don't, so I have to do `null as any` to trick the type system.